requirements.txt
SoundFile
librosa
numpy
scipy
sklearn
pytorch-lightning==1.7.4
matplotlib
seaborn
hydra-core
pyworld
transformers
pysptk
nnmnkwii
seaborn
tensorboard
fairseq
